Refactored code.| #3022
